queries secured by authentication

+++ src/resolvers/Mutation.js
@@ -1,5 +1,6 @@
 import bcrypt from "bcrypt"
 import jwt from "jsonwebtoken"
+import getUserId from "../utils/getUserId"
 
 const Mutation = {
 	async createUser(parent, { data }, { prisma }, info) {
@@ -17,18 +18,22 @@ const Mutation = {
 
 		return { user, token: jwt.sign({ userId: user.id }, "thisisasecret") }
 	},
-	deleteUser(parent, args, { prisma }, info) {
+	deleteUser(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+
 		return prisma.mutation.deleteUser(
 			{
-				where: { id: args.id },
+				where: { id: userId },
 			},
 			info
 		)
 	},
-	updateUser(parent, { data, id }, { prisma }, info) {
+	updateUser(parent, { data }, { prisma, request }, info) {
+		const userId = getUserId(request)
+
 		return prisma.mutation.updateUser(
 			{
-				where: { id },
+				where: { id: userId },
 				data,
 			},
 			info
@@ -36,7 +41,7 @@ const Mutation = {
 	},
 	async login(parent, { data }, { prisma }, info) {
 		const { email, password } = data
-		
+
 		const user = await prisma.query.user({ where: { email } })
 
 		if (!user) throw new Error("Unable to login")
@@ -47,21 +52,31 @@ const Mutation = {
 
 		return { user, token: jwt.sign({ userId: user.id }, "thisisasecret") }
 	},
-	createPost(parent, { data }, { prisma }, info) {
-		const { title, body, published, author } = data
+	createPost(parent, { data }, { prisma, request }, info) {
+		const userId = getUserId(request)
+		const { title, body, published } = data
+
 		return prisma.mutation.createPost(
 			{
 				data: {
 					title,
 					body,
 					published,
-					author: { connect: { id: author } },
+					author: { connect: { id: userId } },
 				},
 			},
 			info
 		)
 	},
-	deletePost(parent, args, { prisma }, info) {
+	async deletePost(parent, args, { prisma, request }, info) {
+		const userId = getUserId(request)
+		const postExists = await prisma.exists.Post({
+			id: args.id,
+			author: { id: userId },
+		})
+
+		if (!postExists) throw new Error("Unable to delete post")
+
 		return prisma.mutation.deletePost(
 			{
 				where: { id: args.id },
@@ -69,7 +84,15 @@ const Mutation = {
 			info
 		)
 	},
-	updatePost(parent, { id, data }, { prisma }, info) {
+	async updatePost(parent, { id, data }, { prisma, request }, info) {
+		const userId = getUserId(request)
+		const postExists = await prisma.exists.Post({
+			id,
+			author: { id: userId },
+		})
+
+		if (!postExists) throw new Error("Unable to update post")
+
 		return prisma.mutation.updatePost(
 			{
 				data,
@@ -78,20 +101,30 @@ const Mutation = {
 			info
 		)
 	},
-	createComment(parent, { data }, { prisma }, info) {
-		const { text, author, post } = data
+	createComment(parent, { data }, { prisma, request }, info) {
+		const userId = getUserId(request)
+		const { text, post } = data
+		
 		return prisma.mutation.createComment(
 			{
 				data: {
 					text,
-					author: { connect: { id: author } },
+					author: { connect: { id: userId } },
 					post: { connect: { id: post } },
 				},
 			},
 			info
 		)
 	},
-	deleteComment(parent, { id }, { prisma }, info) {
+	async deleteComment(parent, { id }, { prisma, request }, info) {
+		const userId = getUserId(request)
+		const commentExists = await prisma.exists.Comment({
+			id,
+			author: { id: userId },
+		})
+
+		if(!commentExists) throw new Error("Unable to delete comment")
+
 		return prisma.mutation.deleteComment(
 			{
 				where: { id },
@@ -99,7 +132,15 @@ const Mutation = {
 			info
 		)
 	},
-	updateComment(parent, { id, data }, { prisma }, info) {
+	async updateComment(parent, { id, data }, { prisma, request }, info) {
+		const userId = getUserId(request)
+		const commentExists = await prisma.exists.Comment({
+			id,
+			author: { id: userId },
+		})
+
+		if(!commentExists) throw new Error("Unable to delete comment")
+
 		return prisma.mutation.updateComment(
 			{
 				data,
